store: Tighten up how WritableStore uses SubgraphStore
The SubgraphStore gives access to all instances of a deployment, but for
the WritableStore we need to be very careful that we do not accidentally
query or modify another deployment instance than the one in the site. These
code changes will hopefully make it more obvious when we would rely on
SubgraphStore functionality that is dependent on the precise deployment
instance.